+++ graphql-maven-plugin-logic/src/main/resources/templates/client_jackson_deserialize.vm.java
@@ -31,7 +31,7 @@ public class CustomScalarDeserializer${object.name}  extends AbstractCustomScala
 #elseif (${object.graphQLScalarTypeStaticField})
 				${object.graphQLScalarTypeStaticField}
 #elseif (${object.graphQLScalarTypeGetter})
-				${object.graphQLScalarTypeGetter}()
+				${object.graphQLScalarTypeGetter}
 #else
 			${object.javaName} : you must define one of graphQLScalarTypeClass, graphQLScalarTypeStaticField or graphQLScalarTypeGetter (in the POM parameters for CustomScalars)
 #end