+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/compilation_tests/AbstractIntegrationTest.java
@@ -1,7 +1,5 @@
 package com.graphql_java_generator.plugin.compilation_tests;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import java.io.IOException;
 
 import javax.annotation.Resource;
@@ -11,9 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.test.annotation.DirtiesContext;
 
-import com.graphql_java_generator.plugin.CodeGenerator;
-import com.graphql_java_generator.plugin.DocumentParser;
 import com.graphql_java_generator.plugin.PluginConfiguration;
+import com.graphql_java_generator.plugin.PluginEntryPoint;
 import com.graphql_java_generator.plugin.test.compiler.CompilationTestHelper;
 import com.graphql_java_generator.plugin.test.helper.GraphqlTestHelper;
 import com.graphql_java_generator.plugin.test.helper.MavenTestHelper;
@@ -32,11 +29,8 @@ abstract class AbstractIntegrationTest {
 
 	@Resource
 	PluginConfiguration pluginConfiguration;
-
-	@javax.annotation.Resource
-	protected DocumentParser documentParser;
-	@javax.annotation.Resource
-	protected CodeGenerator codeGenerator;
+	@Resource
+	PluginEntryPoint pluginEntryPoint;
 
 	/**
 	 * This test will be executed for each concrete subclass of this class
@@ -48,18 +42,11 @@ abstract class AbstractIntegrationTest {
 	@DirtiesContext // We need to forget the previous parsing (or everything may be doubled)
 	void testGenerateCode() throws IOException {
 		// Preparation
-		int i = documentParser.parseDocuments();
-
 		mavenTestHelper.deleteDirectoryAndContentIfExists(pluginConfiguration.getTargetSourceFolder());
 		mavenTestHelper.deleteDirectoryAndContentIfExists(pluginConfiguration.getTargetClassFolder());
 
 		// Go, go, go
-		int verif = codeGenerator.generateCode();
-
-		// Basic verification of the number of generated files. The samples will work only if all needed files are
-		// generated
-		// (checking properly the number is not that simple, and changes to often to maintain it)
-		assertTrue(verif > i, "More file should be generated than what's parsed");
+		pluginEntryPoint.execute();
 
 		compilationTestHelper.checkCompleteCompilationStatus(null);
 	}