+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/SpringConfiguration.java
@@ -3,11 +3,6 @@
  */
 package com.graphql_java_generator.mavenplugin;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.apache.maven.plugin.MojoExecutionException;
 import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -15,10 +10,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 import com.graphql_java_generator.plugin.PluginConfiguration;
-import com.graphql_java_generator.plugin.ResourceSchemaStringProvider;
-
-import graphql.language.Document;
-import graphql.parser.Parser;
 
 /**
  * @author EtienneSF
@@ -37,26 +28,5 @@ public class SpringConfiguration {
 	@Bean
 	PluginConfiguration pluginConfiguration() {
 		return new PluginConfigurationImpl(mojo);
-	}
-
-	/**
-	 * Loads the schema from the graphqls files. This method uses the {@link GraphQLJavaToolsAutoConfiguration} from the
-	 * 
-	 * project, to load the schema from the graphqls files
-	 * 
-	 * @throws MojoExecutionException
-	 *             When an error occurs while reading or parsing the graphql definition files
-	 */
-	@Bean
-	public List<Document> documents(ResourceSchemaStringProvider schemaStringProvider) throws MojoExecutionException {
-		try {
-			Parser parser = new Parser();
-			return schemaStringProvider.schemaStrings().stream().map(parser::parseDocument)
-					.collect(Collectors.toList());
-		} catch (IOException e) {
-			throw new MojoExecutionException("Error while reading graphql schema definition files: " + e.getMessage(),
-					e);
-		}
-
 	}
 }